OcrServer/requirements.txt
numpy
cnocr
Pillow
Django
